requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pydicom==2.4.3
Pillow==10.0.0
numpy==1.24.3
reportlab==4.0.4
pystray~=0.19.5
pyinstaller~=6.17.0